fix: enforce bridge destination compliance on depositAndBridge
depositAndBridge was only verifying deposit compliance (depositor,
asset, amount) but not the bridge destination address. The standalone
bridge() enforced this via _verifyBridgeCompliance, creating a bypass
path where a compliant depositor could bridge shares to a non-compliant
address.

Add _verifyDepositAndBridgeCompliance to CrossChainTellerWithGenericBridge
that hashes the bridge `to` into the compliance message, so the signer
explicitly approves the full deposit-and-bridge action in one signature.

+++ src/base/Roles/AccountantWithYieldStreaming.sol
@@ -105,8 +105,8 @@ contract AccountantWithYieldStreaming is AccountantWithRateProviders {
     event ExchangeRateUpdated(uint256 newExchangeRate);
     event MaximumVestDurationUpdated(uint64 newMaximum);
     event MinimumVestDurationUpdated(uint64 newMinimum);
-    event MaximumDeviationYieldUpdated(uint64 newMaximum);
-    event MaximumDeviationLossUpdated(uint64 newMaximum);
+    event MaximumDeviationYieldUpdated(uint32 newMaximum);
+    event MaximumDeviationLossUpdated(uint32 newMaximum);
 
     constructor(
         address _owner,
